add and fetch post likers

+++ database/migrations/2019_08_27_172508_create_posts_table.php
@@ -14,8 +14,8 @@ class CreatePostsTable extends Migration
     public function up()
     {
         Schema::create('posts', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('user_id')->unsigned()->index();
+            $table->increments('id');
+            $table->integer('user_id')->unsigned()->index();
             $table->text('body');
             $table->timestamps();
